Add files via upload
Xpert Learning Assistant aided me in the use of the format x.Cells(y.Row, 1) function to point to the same row as the current selection of a different column, we did not discuss this specifically in class I don't recall, not the .Row function in this regard. Stating as to honor code source. Thank you for your time and review!